+++ src/webview/vscode.d.ts
@@ -24,3 +24,4 @@ declare const partyUri: string;
 declare const hiddenStatuses: string[];
 declare const language: string;
 declare const showTips: boolean;
+declare const fontFamily: string;